..unde jetzt tut's auch richtig, ohne in der falschen main() zu hängen

+++ aodvv2/seqnum.h
@@ -6,6 +6,6 @@
 void seqNum_init(void);
 uint16_t seqNum_get(void);
 void seqNum_inc(void);
-int seqnum_cmp(uint32_t s1, uint32_t s2);
+int seqnum_cmp(uint16_t s1, uint16_t s2);
 
 #endif /* SEQNUM_H_ */
